Support the simplest use case
The tool is intended to erase a subset of coverage statements generated
after the execution of `go tool cover` over a source file. The subset of
erased statements are chosen following this criteria:
    - For each `ast.BlockStmt`, just keep the last statement in the
      `List` associated to it.
    - Switch statements are a special case, because `go tool cover`
      instrumentates the body of each `ast.CaseStmt`.
The particular use case of this, is to erase the instrumentation entries
whose state can be inferred by using information about other
instrumentation entries, based on the execution paths the program can
take.

This adds support to the simplest use case by always deleting the first
statement in a `ast.BlockStmt`. The simplest use case looks like:

```
func Foo() {<InstrumentationEntry1>
    ...
    if <condition> {<InstrumentationEntry3>
        ...
    }
    <InstrumentationEntry2>;...
}
```
In this case, `<InstrumentationEntry1>` is deleted.